Add OpenTelemetry tracing
This PR adds support for OpenTelemetry based tracing, with automatic
links between job creation, and processing.

It also adds instrumentation to most db functions, except get_job, as
that one seemed to be called a lot, so it felt a bit excessive.

I've tried to follow the related OTel semantic conventions, which seems
to be these:

https://opentelemetry.io/docs/specs/semconv/messaging/messaging-spans/

But I'm open for using other conventions, and/or change which tags are
exposed.