feat: re-implement queue names with a lazy-loading queue cache
This gives us the queue name in our lifecycle hooks without needing a
sql query per job lifecycle moment. Instead, we follow the pattern
TaskDetails establishes: we keep a hashmap of queues we've seen with
names.

This cache is unlike TaskDetails in that it needs interior mutability to
update itself through a worker's life-- queues are created on the fly in
a postgres stored procedure when needed. We lazy-load and query only on
a cache miss.

The name is still an Option in the hook data, because some queues don't
have names, so we remain backwards compatible.

Implementation notes:
- added QueueDetails, following the TaskDetails example
- QueueDetails uses an Arc<RwLock<HashMap>> for thread-safe updates
- it also has a PgPool so it can make queries
- updated all hook call sites to use async get() method

Performance notes:
- This is definitely better than querying on every lifecycle hook.
- Cache hits take ~100ns, cache misses trigger one DB query per
  unique queue, which probably is not a lot in a typical deployment
  (Mine will be around 10 tops.)
- This adds a little bit of memory use per queue.