feat: add lifecycle events for JobCreating and JobCreated
Creating is fired during creation, before database insert. Created is
fired after successful job creation and storage. Both add_job() and
add_raw_job() fire the hooks for these events. To send them, we had to
make the list of hooks available to the worker utils via a reference.

The JobCreating and JobStarted event structures now both have the job
payload data, as requested to support the tracing PR. In both cases the
payload was available already to be cloned into the event structure.

Implementation notes:
- updated tests & documentation
- renamed JobLifeCycleEvent to LifeCycleEvent for potential future
  generalization-- other events might be wanted!
- added enum variants Creating and Created
- create_utils() is now implemented in terms of a new
  create_utils_with_hooks() function
- similarly, WorkerUtils::new() calls WorkerUtils::new_with_hooks()
  with an empty list

This addresses the request for support in the hooks PR to potentially
allow tracing to be implemented via hooks.